fix failed tests

+++ account/signals.py
@@ -5,7 +5,8 @@ from account.models import Project, CustomUser, UserBankAccount
 
 @receiver([post_save, post_delete], sender=Project)
 def invalidate_project_cache(sender, instance, **kwargs):
-    cache.delete_pattern(f"*task_completion_stats_{instance.created_by.id}*")
+    if instance.created_by:
+        cache.delete_pattern(f"*task_completion_stats_{instance.created_by.id}*")
     cache.delete_pattern(f"*project_detail_GET_/api/v1/account/projects/{instance.id}/*")
     
 @receiver([post_save, post_delete], sender=CustomUser)
